Moving 'opsgenie' code into the modules directory for 3rd party libs
OpsGenieAlerter moved to modules/opsgenie.py to separate 3rd party
additions from the core systems.
